Do not allow to delete base roles and update Admin role

+++ src/routes/role/role.service.ts
@@ -1,5 +1,5 @@
 import { BadRequestException, Injectable } from '@nestjs/common'
-import { AlreadyExistedRoleException } from 'src/routes/role/role.error'
+import { AlreadyExistedRoleException, ProhibitedActionOnBaseRoleException } from 'src/routes/role/role.error'
 import {
   CreateRoleRequestBodyType,
   CreateRoleResponseType,
@@ -9,9 +9,11 @@ import {
   UpdateRoleRequestBodyType,
 } from 'src/routes/role/role.model'
 import { RoleRepository } from 'src/routes/role/role.repo'
+import { RoleName } from 'src/shared/constants/role.constant'
 import { NotFoundRecordException } from 'src/shared/error'
 import { isPrismaNotFoundError, isPrismaUniqueConstraintFailedError } from 'src/shared/helpers'
 import { ResponseMessageType } from 'src/shared/models/response.model'
+import { keyof } from 'zod'
 
 @Injectable()
 export class RoleService {
@@ -43,6 +45,14 @@ export class RoleService {
     updatedByUserId: number
   }): Promise<RoleDetailsType> {
     try {
+      const role = await this.roleRepository.findById(payload.id)
+      if (!role) throw NotFoundRecordException
+
+      // MEMO: Không cho phép UPDATE role Admin
+      if (role.name === RoleName.Admin) {
+        throw ProhibitedActionOnBaseRoleException
+      }
+
       const roleWithPermissions = await this.roleRepository.update(payload)
       return roleWithPermissions
     } catch (error) {
@@ -55,6 +65,15 @@ export class RoleService {
 
   async delete(payload: { id: number; updatedByUserId: number }): Promise<ResponseMessageType> {
     try {
+      const role = await this.roleRepository.findById(payload.id)
+      if (!role) throw NotFoundRecordException
+
+      // MEMO: Không cho phép DELETE các role cơ bản
+      const baseRoles: string[] = [RoleName.Admin, RoleName.Seller, RoleName.Client]
+      if (baseRoles.includes(role.name)) {
+        throw ProhibitedActionOnBaseRoleException
+      }
+
       await this.roleRepository.delete(payload)
       return { message: 'Delete role successfully' }
     } catch (error) {
